ollama_web_search/pyproject.toml
[project]
name = "ollama-web-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gpt-oss>=0.0.8",
    "ollama>=0.6.0",
    "python-dotenv>=1.1.1",
]
